feat: add NGWMN getters as an `ogc` sibling; extract a shared OGC engine
Port the NGWMN functions from the R `dataRetrieval` package
(DOI-USGS/dataRetrieval#904) and refactor the Water Data OGC machinery into a
generic, API-agnostic engine, so NGWMN and Water Data are sibling layers on
top of it -- NGWMN does not depend on Water Data.

dataretrieval/ogc/  generic OGC engine (no service-specific config)
  engine.py    request build, pagination, parse/finalize, get_ogc_data
  chunking.py  URL-byte multi-value chunker
  filters.py   CQL `filter` splitting
  progress.py  self-updating status line
The engine is parameterized by an `OgcDialect` and a base-url context
variable rather than branching on service names: Water Data POSTs CQL2 for
`monitoring-locations` and renders `daily` time args date-only; NGWMN needs
neither. Adding a sibling API is a new dialect + base URL, not engine edits.

dataretrieval/ngwmn.py  sibling getters that import only dataretrieval.ogc:
  get_sites, get_water_level, get_lithology, get_well_construction,
  get_providers

dataretrieval/waterdata/  thin Water Data layer on the engine; the Statistics
API lives in its own waterdata/stats.py module.

Unified `state` parameter across the modern getters, accepting a full name, a
two-letter postal code, or a two-digit ANSI/FIPS code; normalized by
codes.states.to_state (50 states + DC, fails fast on a typo) and resolved at
the getter layer. The native state_code/state_name parameters remain as an
escape hatch.

Also: export ChunkInterrupted at the package top level; key the empty-result
short-circuit off `features` (NGWMN omits `numberReturned`) and tolerate
geometry-less features; always return PEP-8 snake_case columns; and add a
pre-commit mypy hook mirroring the CI type-check.
