+++ dataretrieval/__init__.py
@@ -19,7 +19,9 @@ imported on demand: ``from dataretrieval import nldi``.
 
 A failed request raises a subclass of :class:`dataretrieval.DataRetrievalError`
 (the taxonomy lives in ``dataretrieval.exceptions``); connection-level failures
-(timeouts, DNS) are wrapped as :class:`dataretrieval.NetworkError`.
+(timeouts, DNS) are wrapped as :class:`dataretrieval.NetworkError`. A large
+request interrupted mid-stream raises :class:`dataretrieval.ChunkInterrupted`,
+whose ``.call.resume()`` continues from the work already completed.
 """
 
 from importlib.metadata import PackageNotFoundError, version
@@ -42,9 +44,22 @@ from dataretrieval.exceptions import (
     URLTooLong,
 )
 
+# Resumable chunk-interruption exceptions. They are defined in
+# ``dataretrieval.ogc.chunking`` rather than ``dataretrieval.exceptions``
+# because they carry pandas/httpx state and a resumable ``ChunkedCall`` handle,
+# which would pull heavy dependencies into the lightweight exceptions module.
+# Surfaced here so callers get a stable public path:
+# ``from dataretrieval import ChunkInterrupted``.
+from dataretrieval.ogc.chunking import (
+    ChunkInterrupted,
+    QuotaExhausted,
+    ServiceInterrupted,
+)
+
 from . import (
     exceptions,
     nadp,
+    ngwmn,
     nwis,
     samples,
     streamstats,
@@ -56,6 +71,7 @@ from . import (
 __all__ = [
     # service modules
     "nadp",
+    "ngwmn",
     "nwis",
     "samples",
     "streamstats",
@@ -75,5 +91,9 @@ __all__ = [
     "TransientError",
     "URLTooLong",
     "Unchunkable",
+    # resumable chunk-interruption exceptions (defined in ogc.chunking)
+    "ChunkInterrupted",
+    "QuotaExhausted",
+    "ServiceInterrupted",
     "__version__",
 ]
