+++ dataretrieval/ogc/__init__.py
@@ -1,0 +1,1 @@
+"""Generic OGC API engine shared by the Water Data and NGWMN getters."""
