+++ dataretrieval/waterdata/__init__.py
@@ -9,6 +9,8 @@ See https://api.waterdata.usgs.gov/ for API reference.
 
 from __future__ import annotations
 
+from dataretrieval.ogc.filters import FILTER_LANG
+
 # Public API exports
 from .api import (
     get_channel,
@@ -30,7 +32,6 @@ from .api import (
     get_stats_por,
     get_time_series_metadata,
 )
-from .filters import FILTER_LANG
 from .nearest import get_nearest_continuous
 from .ratings import get_ratings
 from .types import (
