+++ dataretrieval/waterdata/api.py
@@ -16,6 +16,7 @@ from urllib.parse import quote
 import httpx
 import pandas as pd
 
+from dataretrieval.ogc.filters import FILTER_LANG
 from dataretrieval.utils import (
     HTTPX_DEFAULTS,
     BaseMetadata,
@@ -23,7 +24,7 @@ from dataretrieval.utils import (
     _get,
     to_str,
 )
-from dataretrieval.waterdata.filters import FILTER_LANG
+from dataretrieval.waterdata import stats
 from dataretrieval.waterdata.types import (
     CODE_SERVICES,
     METADATA_COLLECTIONS,
@@ -45,8 +46,8 @@ from dataretrieval.waterdata.utils import (
     _run_sync,
     _switch_properties_id,
     _walk_pages,
+    _with_state,
     get_ogc_data,
-    get_stats_data,
 )
 
 # Set up logger for this module
@@ -200,7 +201,7 @@ def get_daily(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -252,7 +253,7 @@ def get_daily(
         >>> # multiple sub-requests so the URL stays under the server's byte
         >>> # limit. Combined output looks like a single query.
         >>> sites_df, _ = dataretrieval.waterdata.get_monitoring_locations(
-        ...     state_name="Ohio",
+        ...     state="Ohio",
         ...     site_type="Stream",
         ... )
         >>> df, md = dataretrieval.waterdata.get_daily(
@@ -408,7 +409,7 @@ def get_continuous(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -433,7 +434,7 @@ def get_continuous(
         ... )
 
         >>> # Pull several disjoint time windows in one call via a CQL
-        >>> # ``filter``. See ``dataretrieval.waterdata.filters`` for the
+        >>> # ``filter``. See ``dataretrieval.ogc.filters`` for the
         >>> # full grammar, auto-chunking, and pitfalls.
         >>> df, md = dataretrieval.waterdata.get_continuous(
         ...     monitoring_location_id="USGS-02238500",
@@ -464,6 +465,7 @@ def get_monitoring_locations(
     district_code: str | Iterable[str] | None = None,
     country_code: str | Iterable[str] | None = None,
     country_name: str | Iterable[str] | None = None,
+    state: str | Iterable[str] | None = None,
     state_code: str | Iterable[str] | None = None,
     state_name: str | Iterable[str] | None = None,
     county_code: str | Iterable[str] | None = None,
@@ -545,6 +547,10 @@ def get_monitoring_locations(
         The code for the country in which the monitoring location is located.
     country_name : string or iterable of strings, optional
         The name of the country in which the monitoring location is located.
+    state : string or iterable of strings, optional
+        State/territory filter (the recommended parameter). Accepts a full name
+        (``"Wisconsin"``), a two-letter postal code (``"WI"``), or a two-digit
+        ANSI/FIPS code (``"55"``).
     state_code : string or iterable of strings, optional
         State code. A two-digit ANSI code (formerly FIPS code) as defined by
         the American National Standards Institute, to define States and
@@ -713,7 +719,7 @@ def get_monitoring_locations(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -744,8 +750,9 @@ def get_monitoring_locations(
     """
     service = "monitoring-locations"
 
-    # Build argument dictionary, omitting None values
-    args = _get_args(locals())
+    # Build argument dictionary, omitting None values (resolving the unified
+    # `state` argument into the OGC `state_name` queryable).
+    args = _get_args(_with_state(locals(), to="name", into="state_name"))
 
     return get_ogc_data(args, service)
 
@@ -757,6 +764,7 @@ def get_time_series_metadata(
     properties: str | Iterable[str] | None = None,
     statistic_id: str | Iterable[str] | None = None,
     hydrologic_unit_code: str | Iterable[str] | None = None,
+    state: str | Iterable[str] | None = None,
     state_name: str | Iterable[str] | None = None,
     last_modified: str | Iterable[str] | None = None,
     begin: str | Iterable[str] | None = None,
@@ -823,6 +831,10 @@ def get_time_series_metadata(
         to the largest (regions). Each hydrologic unit is identified by a unique
         hydrologic unit code (HUC) consisting of two to eight digits based on the
         four levels of classification in the hydrologic unit system.
+    state : string or iterable of strings, optional
+        State/territory filter (the recommended parameter). Accepts a full name
+        (``"Wisconsin"``), a two-letter postal code (``"WI"``), or a two-digit
+        ANSI/FIPS code (``"55"``).
     state_name : string or iterable of strings, optional
         The name of the state or state equivalent in which the monitoring location
         is located.
@@ -937,7 +949,7 @@ def get_time_series_metadata(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -968,8 +980,9 @@ def get_time_series_metadata(
     """
     service = "time-series-metadata"
 
-    # Build argument dictionary, omitting None values
-    args = _get_args(locals())
+    # Build argument dictionary, omitting None values (resolving the unified
+    # `state` argument into the OGC `state_name` queryable).
+    args = _get_args(_with_state(locals(), to="name", into="state_name"))
 
     return get_ogc_data(args, service)
 
@@ -998,6 +1011,7 @@ def get_combined_metadata(
     district_code: str | Iterable[str] | None = None,
     country_code: str | Iterable[str] | None = None,
     country_name: str | Iterable[str] | None = None,
+    state: str | Iterable[str] | None = None,
     state_code: str | Iterable[str] | None = None,
     state_name: str | Iterable[str] | None = None,
     county_code: str | Iterable[str] | None = None,
@@ -1106,6 +1120,10 @@ def get_combined_metadata(
         interval (``"start/end"``, optionally half-bounded with ``..``),
         or an ISO 8601 duration (e.g. ``"P1M"``, ``"PT36H"``). See
         :func:`get_time_series_metadata` for the full grammar.
+    state : string or iterable of strings, optional
+        State/territory filter (the recommended parameter). Accepts a full
+        name (``"Wisconsin"``), a two-letter postal code (``"WI"``), or a
+        two-digit ANSI/FIPS code (``"55"``).
     state_name, county_name, hydrologic_unit_code, site_type, \
 site_type_code : string or iterable of strings, optional
         Common location-catalog filters carried over from the
@@ -1131,7 +1149,7 @@ site_type_code : string or iterable of strings, optional
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -1160,7 +1178,7 @@ site_type_code : string or iterable of strings, optional
 
         >>> # Every series in a single county, useful for area-of-interest workflows
         >>> df, md = dataretrieval.waterdata.get_combined_metadata(
-        ...     state_name="Wisconsin", county_name="Dane County"
+        ...     state="Wisconsin", county_name="Dane County"
         ... )
 
         >>> # Inventory across multiple HUCs, restricted to streams and springs
@@ -1198,7 +1216,8 @@ site_type_code : string or iterable of strings, optional
     """
     service = "combined-metadata"
 
-    args = _get_args(locals())
+    # Resolve the unified `state` argument into the OGC `state_name` queryable.
+    args = _get_args(_with_state(locals(), to="name", into="state_name"))
 
     return get_ogc_data(args, service)
 
@@ -1347,7 +1366,7 @@ def get_latest_continuous(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -1543,7 +1562,7 @@ def get_latest_daily(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -1731,7 +1750,7 @@ def get_field_measurements(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -1846,7 +1865,7 @@ def get_field_measurements_metadata(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -1969,7 +1988,7 @@ def get_peaks(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
@@ -2443,6 +2462,7 @@ def get_stats_por(
     approval_status: str | None = None,
     computation_type: str | Iterable[str] | None = None,
     country_code: str | Iterable[str] | None = None,
+    state: str | Iterable[str] | None = None,
     state_code: str | Iterable[str] | None = None,
     county_code: str | Iterable[str] | None = None,
     start_date: str | None = None,
@@ -2453,6 +2473,7 @@ def get_stats_por(
     site_type_code: str | Iterable[str] | None = None,
     site_type_name: str | Iterable[str] | None = None,
     parameter_code: str | Iterable[str] | None = None,
+    normal_type: str | None = None,
     expand_percentiles: bool = True,
 ) -> tuple[pd.DataFrame, BaseMetadata]:
     """Get day-of-year and month-of-year water data statistics from the
@@ -2478,6 +2499,10 @@ def get_stats_por(
         arithmetic_mean, maximum, median, minimum, percentile.
     country_code: string, optional
         Country query parameter. API defaults to "US".
+    state: string or iterable of strings, optional
+        State/territory filter (the recommended parameter). Accepts a full name
+        ("Wisconsin"), a two-letter postal code ("WI"), or a two-digit
+        ANSI/FIPS code ("55").
     state_code: string, optional
         State query parameter. Takes the format "US:XX", where XX is
         the two-digit state code. API defaults to "US:42" (Pennsylvania).
@@ -2514,6 +2539,10 @@ def get_stats_por(
         measured and the units of measure. A complete list of parameter codes
         and associated groupings can be found at
         https://help.waterdata.usgs.gov/codes-and-parameters/parameters.
+    normal_type : string, optional
+        Filter the returned normals to a single period. If unspecified
+        (default), all matching data are returned. Available values:
+        "DOY" (day-of-year) and "MOY" (month-of-year).
     expand_percentiles : boolean
         Percentile data for a given day of year or month of year by default
         are returned from the service as lists of string values and percentile
@@ -2563,9 +2592,12 @@ def get_stats_por(
         ... )
     """
     # Build argument dictionary, omitting None values
-    params = _get_args(locals(), exclude={"expand_percentiles"})
+    params = _get_args(
+        _with_state(locals(), to="fips_us", into="state_code"),
+        exclude={"expand_percentiles"},
+    )
 
-    return get_stats_data(
+    return stats.get_data(
         args=params, service="observationNormals", expand_percentiles=expand_percentiles
     )
 
@@ -2574,6 +2606,7 @@ def get_stats_date_range(
     approval_status: str | None = None,
     computation_type: str | Iterable[str] | None = None,
     country_code: str | Iterable[str] | None = None,
+    state: str | Iterable[str] | None = None,
     state_code: str | Iterable[str] | None = None,
     county_code: str | Iterable[str] | None = None,
     start_date: str | None = None,
@@ -2584,6 +2617,7 @@ def get_stats_date_range(
     site_type_code: str | Iterable[str] | None = None,
     site_type_name: str | Iterable[str] | None = None,
     parameter_code: str | Iterable[str] | None = None,
+    interval_type: str | Iterable[str] | None = None,
     expand_percentiles: bool = True,
 ) -> tuple[pd.DataFrame, BaseMetadata]:
     """Get monthly and annual water data statistics from the USGS Water Data API.
@@ -2608,6 +2642,10 @@ def get_stats_date_range(
         arithmetic_mean, maximum, median, minimum, percentile.
     country_code: string, optional
         Country query parameter. API defaults to "US".
+    state: string or iterable of strings, optional
+        State/territory filter (the recommended parameter). Accepts a full name
+        ("Wisconsin"), a two-letter postal code ("WI"), or a two-digit
+        ANSI/FIPS code ("55").
     state_code: string, optional
         State query parameter. Takes the format "US:XX", where XX is
         the two-digit state code. API defaults to "US:42" (Pennsylvania).
@@ -2649,6 +2687,10 @@ def get_stats_date_range(
         measured and the units of measure. A complete list of parameter codes
         and associated groupings can be found at
         https://help.waterdata.usgs.gov/codes-and-parameters/parameters.
+    interval_type : string or iterable of strings, optional
+        Filter the returned intervals to one or more periods. If unspecified
+        (default), all matching data are returned. Available values:
+        "M" (month), "CY" (calendar year), and "WY" (water year).
     expand_percentiles : boolean
         Percentile data for a given day of year or month of year by default
         are returned from the service as lists of string values and percentile
@@ -2682,7 +2724,7 @@ def get_stats_date_range(
         >>> # Get monthly and yearly medians for streamflow at streams in Rhode Island
         >>> # from calendar year 2024.
         >>> df, md = dataretrieval.waterdata.get_stats_date_range(
-        ...     state_code="US:44",  # State code for Rhode Island
+        ...     state="RI",  # Rhode Island (postal code, name, or FIPS all work)
         ...     parameter_code="00060",
         ...     site_type_code="ST",
         ...     start_date="2024-01-01",
@@ -2699,9 +2741,12 @@ def get_stats_date_range(
         ... )
     """
     # Build argument dictionary, omitting None values
-    params = _get_args(locals(), exclude={"expand_percentiles"})
+    params = _get_args(
+        _with_state(locals(), to="fips_us", into="state_code"),
+        exclude={"expand_percentiles"},
+    )
 
-    return get_stats_data(
+    return stats.get_data(
         args=params,
         service="observationIntervals",
         expand_percentiles=expand_percentiles,
@@ -2865,7 +2910,7 @@ def get_channel(
     filter, filter_lang : optional
         Server-side CQL filter passed through as the OGC ``filter`` /
         ``filter-lang`` query parameters. See
-        :mod:`dataretrieval.waterdata.filters` for syntax, auto-chunking,
+        :mod:`dataretrieval.ogc.filters` for syntax, auto-chunking,
         and the lexicographic-comparison pitfall.
     convert_type : boolean, optional
         If True, converts columns to appropriate types.
