+++ docs/source/examples/USGS_NGWMN_Examples.nblink
@@ -1,0 +1,3 @@
+{
+    "path": "../../../demos/USGS_NGWMN_Examples.ipynb"
+}
