+++ docs/source/reference/ngwmn.rst
@@ -1,0 +1,8 @@
+.. _ngwmn:
+
+dataretrieval.ngwmn
+-------------------
+
+.. automodule:: dataretrieval.ngwmn
+    :members:
+    :special-members:
