+++ tests/waterdata_chunking_test.py
@@ -1,4 +1,4 @@
-"""Tests for ``dataretrieval.waterdata.chunking``.
+"""Tests for ``dataretrieval.ogc.chunking``.
 
 These tests exercise the joint planner with a fake ``build_request``
 whose URL byte length is a deterministic function of its inputs:
@@ -17,6 +17,7 @@ and then fail in production.
 
 import asyncio
 import concurrent.futures
+import contextvars
 import datetime
 import http.server
 import sys
@@ -35,10 +36,8 @@ if sys.version_info < (3, 10):
     pytest.skip("Skip entire module on Python < 3.10", allow_module_level=True)
 
 from dataretrieval.exceptions import DataRetrievalError
-from dataretrieval.utils import HTTPX_DEFAULTS
-from dataretrieval.waterdata import chunking as _chunking
-from dataretrieval.waterdata import utils as _utils
-from dataretrieval.waterdata.chunking import (
+from dataretrieval.ogc import chunking as _chunking
+from dataretrieval.ogc.chunking import (
     _LIST_SEP,
     _NEVER_CHUNK,
     _OR_SEP,
@@ -63,6 +62,8 @@ from dataretrieval.waterdata.chunking import (
     get_active_client,
     multi_value_chunked,
 )
+from dataretrieval.utils import HTTPX_DEFAULTS
+from dataretrieval.waterdata import utils as _utils
 from dataretrieval.waterdata.utils import _DATE_RANGE_PARAMS, _construct_api_requests
 
 
@@ -350,7 +351,7 @@ def test_multi_value_chunked_emits_3d_cartesian_product():
 
 
 def test_multi_value_chunked_lazy_url_limit(monkeypatch):
-    """``url_limit=None`` → resolve chunking._WATERDATA_URL_BYTE_LIMIT at call
+    """``url_limit=None`` → resolve chunking._OGC_URL_BYTE_LIMIT at call
     time, so tests that patch the constant affect this decorator too."""
     calls = []
 
@@ -361,7 +362,7 @@ def test_multi_value_chunked_lazy_url_limit(monkeypatch):
             elapsed=datetime.timedelta(seconds=0.1), headers={}
         )
 
-    monkeypatch.setattr(_chunking, "_WATERDATA_URL_BYTE_LIMIT", 240)
+    monkeypatch.setattr(_chunking, "_OGC_URL_BYTE_LIMIT", 240)
     # 4 sites of 10 chars → exceeds 240 → planner splits.
     fetch({"sites": ["S" * 10 + str(i) for i in range(4)]})
     assert len(calls) > 1, "patched constant should drive chunking"
@@ -657,6 +658,55 @@ def test_resume_produces_dataset_identical_to_uninterrupted_run():
 
     # And every original site must be present exactly once.
     assert sorted(df_a["id"].tolist()) == sorted(sites)
+
+
+def test_resume_rebuilds_in_captured_context():
+    """Regression: sub-requests are rebuilt by reading ambient ContextVars
+    (the engine threads base URL / dialect / row cap that way). A
+    ``call.resume()`` fired AFTER the originating ``with`` block exits —
+    the documented recovery for a mid-stream 429 — must still observe the
+    values active when the call was *created*, not the process defaults.
+    ``ChunkedCall`` snapshots the context at construction and runs every
+    drive inside it; without that snapshot a resumed NGWMN call would
+    rebuild its sub-requests against the wrong (default Water Data) base."""
+    var = contextvars.ContextVar("ctx_probe", default="DEFAULT")
+    observed: list[str] = []
+
+    state = {"calls": 0, "tripped": False}
+
+    async def fetch(args):
+        state["calls"] += 1
+        # The value visible at (re)build time — what _construct_api_requests
+        # would read from _ogc_base_url_var / _dialect_var in production.
+        observed.append(var.get())
+        if state["calls"] == 3 and not state["tripped"]:
+            state["tripped"] = True
+            raise RateLimited("429: Too many requests made.")
+        sites = list(args["sites"])
+        return (pd.DataFrame({"id": sites}), _quota_response(500))
+
+    sites = ["S" * 10 + str(i) for i in range(16)]
+    decorated = multi_value_chunked(build_request=_fake_build, url_limit=240)(fetch)
+
+    # Create + drive the call INSIDE the context, so the snapshot captures "IN".
+    token = var.set("IN")
+    try:
+        with pytest.raises(QuotaExhausted) as excinfo:
+            decorated({"sites": sites})
+    finally:
+        var.reset(token)
+
+    # The originating context has exited — the bare var is back to default.
+    assert var.get() == "DEFAULT"
+    assert 0 < excinfo.value.completed_chunks < excinfo.value.total_chunks
+
+    # Resume OUTSIDE the context. Every rebuilt sub-request must still see
+    # "IN" (the captured snapshot), never the leaked "DEFAULT".
+    observed.clear()
+    df, _ = excinfo.value.call.resume()
+    assert observed, "resume issued no sub-requests"
+    assert set(observed) == {"IN"}, observed
+    assert sorted(df["id"].tolist()) == sorted(sites)
 
 
 def test_chunker_passes_through_non_429_runtime_error():
@@ -994,7 +1044,7 @@ def test_combine_chunk_responses_returns_independent_headers():
 
 def test_paginate_terminates_on_empty_string_cursor():
     """``_paginate``'s loop predicate is ``while cursor is not None``.
-    Parse-response wrappers in ``_walk_pages`` / ``get_stats_data``
+    Parse-response wrappers in ``_walk_pages`` / ``stats.get_data``
     coerce falsy non-None values to None so an empty-string next-
     cursor (a real-but-unusual end-of-stream sentinel some pagination
     APIs use) doesn't trap us in an infinite ``follow_up('')`` loop."""
