+++ tests/waterdata_utils_test.py
@@ -8,9 +8,13 @@ import httpx
 import pandas as pd
 import pytest
 
+import dataretrieval.ogc.engine as _engine_module
+import dataretrieval.waterdata.stats as _stats_module
 import dataretrieval.waterdata.utils as _utils_module
 from dataretrieval.exceptions import DataRetrievalError, HTTPError, TransientError
-from dataretrieval.waterdata.chunking import RateLimited, ServiceUnavailable
+from dataretrieval.ogc.chunking import RateLimited, ServiceUnavailable
+from dataretrieval.waterdata import get_stats_date_range, get_stats_por
+from dataretrieval.waterdata.stats import _handle_nesting, get_data
 from dataretrieval.waterdata.utils import (
     OGC_API_URL,
     _arrange_cols,
@@ -20,13 +24,12 @@ from dataretrieval.waterdata.utils import (
     _format_api_dates,
     _get_args,
     _get_resp_data,
-    _handle_stats_nesting,
     _next_req_url,
     _parse_retry_after,
     _raise_for_non_200,
     _row_cap,
+    _to_snake_case,
     _walk_pages,
-    get_stats_data,
 )
 
 _LOGGER_NAME = _utils_module.__name__
@@ -332,7 +335,7 @@ def test_get_resp_data_handles_missing_features_key():
     """Regression: a 200 with ``numberReturned > 0`` but no
     ``features`` key (real schema-drift shape) used to crash
     ``_get_resp_data`` with ``KeyError`` — wrapped downstream by
-    ``_paginate`` as a generic transport error. ``_handle_stats_nesting``
+    ``_paginate`` as a generic transport error. ``_handle_nesting``
     was already hardened against this; ``_get_resp_data`` now mirrors
     that defensiveness and returns an empty frame instead."""
     resp = mock.Mock()
@@ -340,6 +343,32 @@ def test_get_resp_data_handles_missing_features_key():
     df = _get_resp_data(resp, geopd=False)
     assert df.empty
     assert isinstance(df, pd.DataFrame)
+
+
+def test_next_req_url_follows_link_without_number_returned():
+    """The NGWMN OGC API omits ``numberReturned`` from its page envelope, so
+    ``_next_req_url`` keys the ``next`` link off ``features`` (mirroring
+    ``_get_resp_data``) rather than that count -- otherwise a page that carries
+    features but no count stops pagination after page 1 and silently truncates
+    every multi-page result. A page that carries features still follows its
+    ``next`` link even when ``numberReturned`` is absent."""
+    resp = mock.MagicMock()
+    resp.url = httpx.URL("https://example.com/page1")
+    body = {
+        # NGWMN shape: features present, NO numberReturned key.
+        "features": [{"id": "1"}],
+        "links": [{"rel": "next", "href": "https://example.com/page2"}],
+    }
+    assert _next_req_url(resp, body=body) == "https://example.com/page2"
+
+
+def test_next_req_url_stops_when_no_features():
+    """A page with no features ends pagination regardless of any stray
+    ``next`` link (and regardless of ``numberReturned``)."""
+    resp = mock.MagicMock()
+    resp.url = httpx.URL("https://example.com/page1")
+    body = {"features": [], "links": [{"rel": "next", "href": "https://x/2"}]}
+    assert _next_req_url(resp, body=body) is None
 
 
 def test_walk_pages_does_not_mutate_initial_response():
@@ -413,15 +442,15 @@ def _stats_initial_ok():
     return resp
 
 
-def _run_get_stats_data_with_failure(failure_resp_or_exc, monkeypatch):
-    """Exercise get_stats_data where the initial response succeeds and the
+def _run_get_data_with_failure(failure_resp_or_exc, monkeypatch):
+    """Exercise get_data where the initial response succeeds and the
     paginated follow-up fails as given. Mirrors _walk_pages_with_failure.
-    `monkeypatch` stubs ``_handle_stats_nesting`` so the synthetic minimal
+    `monkeypatch` stubs ``_handle_nesting`` so the synthetic minimal
     response body doesn't need to parse — these tests only assert on the
     pagination loop's error surfacing."""
     monkeypatch.setattr(
-        _utils_module,
-        "_handle_stats_nesting",
+        _stats_module,
+        "_handle_nesting",
         mock.MagicMock(return_value=pd.DataFrame()),
     )
 
@@ -432,7 +461,7 @@ def _run_get_stats_data_with_failure(failure_resp_or_exc, monkeypatch):
     else:
         mock_client.request.return_value = failure_resp_or_exc
 
-    return get_stats_data(
+    return get_data(
         args={"monitoring_location_id": "USGS-1"},
         service="observationNormals",
         expand_percentiles=False,
@@ -440,14 +469,14 @@ def _run_get_stats_data_with_failure(failure_resp_or_exc, monkeypatch):
     )
 
 
-def test_get_stats_data_raises_on_mid_pagination_failure(monkeypatch):
-    """Wiring smoke: ``get_stats_data`` and ``_walk_pages`` share the
+def test_get_data_raises_on_mid_pagination_failure(monkeypatch):
+    """Wiring smoke: ``get_data`` and ``_walk_pages`` share the
     same ``_paginate`` strategy helper, so error-routing behaviour is
     exercised by the ``_walk_pages`` triplet above. This single
-    ``get_stats_data`` mid-pagination case proves the stats-specific
+    ``get_data`` mid-pagination case proves the stats-specific
     follow-up callback is wired into ``_paginate`` correctly."""
     with pytest.raises(DataRetrievalError, match="Paginated request failed") as excinfo:
-        _run_get_stats_data_with_failure(
+        _run_get_data_with_failure(
             httpx.ConnectError("stats-boom"),
             monkeypatch,
         )
@@ -456,7 +485,7 @@ def test_get_stats_data_raises_on_mid_pagination_failure(monkeypatch):
     assert "stats-boom" in str(excinfo.value)
 
 
-def test_get_stats_data_warning_includes_next_token(caplog, monkeypatch):
+def test_get_data_warning_includes_next_token(caplog, monkeypatch):
     """The pagination-failure warning includes the next_token so operators
     can identify which page in the sequence failed. (Addresses Copilot's
     PR #273 review note: the base URL alone drops cursor context.)"""
@@ -470,17 +499,18 @@ def test_get_stats_data_warning_includes_next_token(caplog, monkeypatch):
     }
 
     with pytest.raises(DataRetrievalError):
-        _run_get_stats_data_with_failure(page2_503, monkeypatch)
+        _run_get_data_with_failure(page2_503, monkeypatch)
 
     warnings_ = [r.getMessage() for r in caplog.records if r.levelno == logging.WARNING]
     # The initial response from _stats_initial_ok carries next=tok2.
     assert any("tok2" in m for m in warnings_), warnings_
 
 
-def test_handle_stats_nesting_tolerates_missing_drop_columns():
-    """If the upstream stats response shape ever changes such that one of
-    the columns we try to drop ("type", "properties.data") is absent, the
-    function should still return a DataFrame instead of raising KeyError.
+def test_handle_nesting_tolerates_missing_drop_columns():
+    """If the upstream stats response shape ever changes such that the nested
+    ``data`` column ``_handle_nesting`` drops is absent, the function should
+    still return a DataFrame instead of raising KeyError (the drop uses
+    ``errors="ignore"``).
     """
     body = {
         "next": None,
@@ -501,38 +531,40 @@ def test_handle_stats_nesting_tolerates_missing_drop_columns():
         ],
     }
 
-    df = _handle_stats_nesting(body, geopd=False)
+    df = _handle_nesting(body, geopd=False)
 
     assert len(df) == 1
     assert df["monitoring_location_id"].iloc[0] == "USGS-12345"
 
 
-def test_handle_stats_nesting_returns_empty_on_empty_features():
+def test_handle_nesting_returns_empty_on_empty_features():
     """A mid-pagination empty page ({\"features\": [], \"next\": <tok>})
     must not crash the downstream merge with
     ``KeyError: 'monitoring_location_id'``. The function short-
     circuits to an empty DataFrame so pagination can continue."""
-    df = _handle_stats_nesting({"features": [], "next": None}, geopd=False)
+    df = _handle_nesting({"features": [], "next": None}, geopd=False)
     assert df.empty
 
 
-def test_handle_stats_nesting_empty_preserves_geopd_type():
+def test_handle_nesting_empty_preserves_geopd_type():
     """When geopandas is available, the empty-features short-circuit
     must return a ``GeoDataFrame`` rather than a plain ``DataFrame``.
     Otherwise a subsequent ``pd.concat([empty, geo_page])`` downgrades
     the final result to a plain ``DataFrame`` and strips geometry/CRS
     — a real regression for geopd-installed users on stats queries
     that hit an empty intermediate page."""
-    # Monkeypatch a stub gpd into the utils module so the test runs
-    # whether or not geopandas is actually installed.
+    # Monkeypatch a stub gpd so the test runs whether or not geopandas is
+    # installed. The empty-page short-circuit delegates to the shared
+    # ``engine._empty_feature_frame``, which resolves ``gpd`` from the engine
+    # namespace — so patch it there, not in the stats module.
     fake_gpd = mock.MagicMock()
 
     class _Sentinel:
         pass
 
     fake_gpd.GeoDataFrame = lambda *a, **kw: _Sentinel()
-    with mock.patch.object(_utils_module, "gpd", fake_gpd, create=True):
-        result = _handle_stats_nesting({"features": []}, geopd=True)
+    with mock.patch.object(_engine_module, "gpd", fake_gpd, create=True):
+        result = _handle_nesting({"features": []}, geopd=True)
     assert isinstance(result, _Sentinel)
 
 
@@ -551,17 +583,19 @@ def test_get_resp_data_empty_preserves_geopd_type():
 
     resp = mock.MagicMock()
     resp.json.return_value = {"numberReturned": 0, "features": [], "links": []}
-    with mock.patch.object(_utils_module, "gpd", fake_gpd, create=True):
+    # ``_get_resp_data`` resolves ``gpd`` from the engine namespace -- patch
+    # it there, not in ``utils``.
+    with mock.patch.object(_engine_module, "gpd", fake_gpd, create=True):
         result = _get_resp_data(resp, geopd=True)
     assert isinstance(result, _Sentinel)
 
 
-def test_handle_stats_nesting_tolerates_missing_features_key():
+def test_handle_nesting_tolerates_missing_features_key():
     """A 200 response with a body that doesn't carry ``features`` at
     all (rare but seen in error envelopes) must also short-circuit
     rather than KeyError before the schema-aware extraction even
     runs."""
-    df = _handle_stats_nesting({}, geopd=False)
+    df = _handle_nesting({}, geopd=False)
     assert df.empty
 
 
@@ -860,3 +894,117 @@ def test_check_ogc_requests_raises_typed_on_5xx(httpx_mock):
     )
     with pytest.raises(ServiceUnavailable):
         _check_ogc_requests(endpoint="daily", req_type="schema")
+
+
+@pytest.mark.parametrize(
+    "name, expected",
+    [
+        ("waterLevelObs", "water_level_obs"),  # camelCase -> snake_case
+        ("monitoring_location_id", "monitoring_location_id"),  # already snake
+        ("value", "value"),  # all-lowercase unchanged
+        ("navd88", "navd88"),  # letter/digit boundary NOT split
+        ("someField", "some_field"),  # simple camelCase
+        ("PascalCase", "pascal_case"),  # leading capital
+        # Runs of capitals are best-effort: only the lower->Upper boundary
+        # before the run is split, so the acronym stays glued to the next word.
+        ("someXMLField", "some_xmlfield"),
+    ],
+)
+def test_to_snake_case(name, expected):
+    assert _to_snake_case(name) == expected
+
+
+def test_get_stats_por_forwards_normal_type(monkeypatch):
+    """``normal_type`` reaches the observationNormals request (parity with R's
+    ``read_waterdata_stats_por``). Guards against the param being dropped from
+    the forwarded args (e.g. accidentally added to ``_get_args``'s exclude)."""
+    captured: dict = {}
+
+    def fake_get_data(args, service, expand_percentiles, client=None):
+        captured.update(args=args, service=service)
+        return pd.DataFrame(), mock.Mock()
+
+    monkeypatch.setattr(_stats_module, "get_data", fake_get_data)
+    get_stats_por(monitoring_location_id="USGS-1", normal_type="MOY")
+    assert captured["service"] == "observationNormals"
+    assert captured["args"].get("normal_type") == "MOY"
+
+
+def test_get_stats_date_range_forwards_interval_type(monkeypatch):
+    """``interval_type`` (multi-value) reaches the observationIntervals request
+    (parity with R's ``read_waterdata_stats_daterange``)."""
+    captured: dict = {}
+
+    def fake_get_data(args, service, expand_percentiles, client=None):
+        captured.update(args=args, service=service)
+        return pd.DataFrame(), mock.Mock()
+
+    monkeypatch.setattr(_stats_module, "get_data", fake_get_data)
+    get_stats_date_range(monitoring_location_id="USGS-1", interval_type=["M", "CY"])
+    assert captured["service"] == "observationIntervals"
+    assert captured["args"].get("interval_type") == ["M", "CY"]
+
+
+def test_with_state_routes_into_native_queryable():
+    """``_with_state`` resolves the canonical ``state`` argument into the
+    endpoint's native queryable (any encoding -> the requested representation)
+    and leaves args without ``state`` untouched."""
+    assert _utils_module._with_state({"state": "WI"}, to="name", into="state_name") == {
+        "state_name": "Wisconsin"
+    }
+    assert _utils_module._with_state(
+        {"state": "Wisconsin"}, to="fips_us", into="state_code"
+    ) == {"state_code": "US:55"}
+    # Multi-value state fans out element-wise.
+    assert _utils_module._with_state(
+        {"state": ["WI", "55"]}, to="name", into="state_name"
+    ) == {"state_name": ["Wisconsin", "Wisconsin"]}
+    # No ``state`` -> mapping returned unchanged.
+    assert _utils_module._with_state(
+        {"state_name": "Ohio"}, to="name", into="state_name"
+    ) == {"state_name": "Ohio"}
+
+
+def test_with_state_conflict_raises():
+    """Passing ``state`` together with a native ``state_code``/``state_name``
+    is ambiguous and raises."""
+    with pytest.raises(ValueError, match="not both"):
+        _utils_module._with_state(
+            {"state": "WI", "state_code": "55"}, to="name", into="state_name"
+        )
+    with pytest.raises(ValueError, match="not both"):
+        _utils_module._with_state(
+            {"state": "WI", "state_name": "Wisconsin"}, to="name", into="state_name"
+        )
+
+
+def test_ogc_getter_resolves_state_at_getter_layer(monkeypatch):
+    """The OGC getters resolve the unified ``state`` into ``state_name``
+    themselves (any encoding), so the shared ``get_ogc_data`` wrapper stays
+    state-agnostic."""
+    import dataretrieval.waterdata.api as _api
+
+    captured: dict = {}
+
+    def fake_get_ogc_data(args, service, *a, **k):
+        captured.update(args=args, service=service)
+        return pd.DataFrame(), mock.Mock()
+
+    monkeypatch.setattr(_api, "get_ogc_data", fake_get_ogc_data)
+    _api.get_monitoring_locations(state="55")  # FIPS in -> full name out
+    assert captured["args"].get("state_name") == "Wisconsin"
+    assert "state" not in captured["args"]
+
+
+def test_get_ogc_data_wrapper_does_not_touch_state():
+    """``get_ogc_data`` no longer rewrites a ``state`` key, so a passthrough
+    query dict (e.g. from ``get_reference_table``) is forwarded untouched."""
+    captured: dict = {}
+
+    def fake_engine_get_ogc_data(args, service, output_id, **k):
+        captured["args"] = dict(args)
+        return pd.DataFrame(), mock.Mock()
+
+    with mock.patch.object(_engine_module, "get_ogc_data", fake_engine_get_ogc_data):
+        _utils_module.get_ogc_data({"state": "WI"}, "monitoring-locations")
+    assert captured["args"] == {"state": "WI"}
